Add support for printing sub-issues
Modified the `Issue` class in `models.py` to include a `children`
attribute that holds a list of sub-issues.

Updated the `gql_fields` method to include the query for retrieving
sub-issues when the `include_children` parameter is set to `True`.

Also added a new `print_issue` function in `print.py` to print sub-issues
recursively when available.